+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom'
+import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { Toaster } from 'sonner'
 import AuthGuard from './components/AuthGuard'
 import LoginPage from './pages/LoginPage'
@@ -16,6 +16,7 @@ import ContestDetailPage from './pages/ContestDetailPage'
 import ContestManagePage from './pages/ContestManagePage'
 import GroupRankingPage from './pages/GroupRankingPage'
 import UserProfilePage from './pages/UserProfilePage'
+import NotFoundPage from './pages/NotFoundPage'
 
 function App() {
   return (
@@ -39,7 +40,7 @@ function App() {
           <Route path="/groups/:groupId/records/manage" element={<GameRecordManagePage />} />
           <Route path="/users/:userId" element={<UserProfilePage />} />
         </Route>
-        <Route path="*" element={<Navigate to="/" replace />} />
+        <Route path="*" element={<NotFoundPage />} />
       </Routes>
     </BrowserRouter>
   )